%laneid: Implement register
The %laneid register returns the lane of the core that a thread is executing
on. From PTX guide: "A predefined, read-only special register that returns
the thread's lane within the warp. The lane identifier ranges from zero to
WARP_SZ-1." Also from the NVIDIA shuffle presentation: "Coordinate of the
thread in a warp: threadIdx.x % 32" [1].

Implement this register.

[1] http://on-demand.gputechconf.com/gtc/2013/presentations/S3174-Kepler-Shuffle-Tips-Tricks.pdf