PTX generation: Use environment ptxas
GPGPU-Sim v3.2 changed to using an environment variable (CUDA_INSTALL_PATH) to
designate the PTX assembler (ptxas) that should be used when generating PTX
code information. Since scons does not source environment variables for
regressions, running gem5-gpu through scons fails to resolve the environment
variable, so ptxas is not found.

In the future, it may make sense to try to fix the environment variable
handling in scons, but for now, revert ptxas path back to older GPGPU-Sim.