shader: Increase core deadlock threshold
Set the core deadlock threshold to 2,000,000 cycles. This is set to be longer
than the deadlock threshold in other parts of the system, because things like
the memory system (ShaderLSQ, Ruby) should indicate deadlock before the shader
core, since accesses originate from activity in the core. This is also
reasonably longer than GPU page faults (<15,000 cycles).