Add GPGPU-Sim 3.2.2 Generated Code
As a step toward making GPGPU-Sim more plug-and-play, move the generated code
from being part of the UW GPGPU-Sim repository and make it part of the patches.
With this change, users can now check out GPGPU-Sim from the UBC git repo,
update it to the correct version, pull UW GPGPU-Sim patches and apply, and
build gem5-gpu.