shader: Round-robin scheduler priority
The default scheduler prioritization often starves higher-index schedulers,
because the low index schedulers are always allowed to issue first. Shuffle the
scheduler priority by round-robining the index of the first scheduler. Note: it
was found empirically that this tends to perform better in the presence of
thread barriers. However, it does not appear that Fermi or Maxwell do this, so
this option is disabled by default.