shader: Parameterize fetch/decode width
Empirical timings show that Maxwell cores must be capable of issuing multiple
independent instructions from a single warp in back-to-back cycles. This is
especially important with the wide issue width of Maxwell (2x that of Fermi).
In order to get reasonable performance comparison from a greedy-then-oldest
warp scheduler in GPGPU-Sim, a warp needs to be able to fetch farther ahead
than just 2 instructions. Parameterize the fetch/decode width.