Local Memory: Implement handling for gem5-gpu
Implement handling of common local memory bits in gem5-gpu. Specifically, allow
local memory accesses to be passed to gem5-gpu CudaCore and ShaderLSQ, add
functionality for libcuda to allocate the local memory region for the GPU, and
map thread-local addresses to the allocated local memory region in the virtual
address space.